feat(FE-Home-page): add edit post functionality

+++ backend/src/controllers/post.controller.ts
@@ -80,7 +80,7 @@ export const getAllPosts = async (
             skip: (Number(page) - 1) * Number(limit),
             limit: Number(limit),
             sort: { [String(sortBy)]: -1 },
-            populate: { path: "author", select: "username" },
+            populate: { path: "author", select: "username profilePicture" },
         };
 
         const posts = await Post.find(query, null, options);
@@ -127,7 +127,10 @@ export const getPostById = async (
 ): Promise<any> => {
     const { postId } = req.params;
     try {
-        const post = await Post.findById(postId).populate("author", "username");
+        const post = await Post.findById(postId).populate(
+            "author",
+            "username profilePicture"
+        );
         if (!post) {
             return res.status(404).json({ error: "Post not found" });
         }
@@ -137,6 +140,7 @@ export const getPostById = async (
         res.status(500).json({ error: "Failed to fetch post" });
     }
 };
+
 // Update a post
 export const updatePost = async (req: Request, res: Response): Promise<any> => {
     const { postId } = req.params;
@@ -145,11 +149,13 @@ export const updatePost = async (req: Request, res: Response): Promise<any> => {
         const updatedPost = await Post.findByIdAndUpdate(
             postId,
             { title, content, tags, status },
-            { new: true }
-        ).populate("author", "username");
+            { new: true } // This option returns the modified document
+        ).populate("author", "username profilePicture");
+
         if (!updatedPost) {
             return res.status(404).json({ error: "Post not found" });
         }
+
         res.status(200).json({
             message: "Post updated successfully",
             post: updatedPost,
@@ -164,6 +170,10 @@ export const updatePost = async (req: Request, res: Response): Promise<any> => {
 export const deletePost = async (req: Request, res: Response): Promise<any> => {
     const { postId } = req.params;
     try {
+        if (!postId) {
+            return res.status(400).json({ error: "Post ID is required" });
+        }
+
         const deletedPost = await Post.findByIdAndDelete(postId);
         if (!deletedPost) {
             return res.status(404).json({ error: "Post not found" });
